최대 주행 거리에 해당하는 자동차 탐색
테스트 완료
